% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbapi-package.R
\docType{package}
\name{lbapi-package}
\alias{lbapi}
\alias{lbapi-package}
\title{lbapi: API wrapper}
\description{
My API wrapper packages including google drive, google sheets and much more.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
