% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs4.R
\name{read_sheet_all}
\alias{read_sheet_all}
\title{Read All Sheets from a Spread Sheet}
\usage{
read_sheet_all(ss, sheets_regex = NULL, ...)
}
\arguments{
\item{ss}{Something that identifies a Google Sheet}

\item{sheets_regex}{If \code{NULL} read all sheets; otherwise, specify regex to subset sheets}

\item{...}{passed to \code{googlesheets4::read_sheet}}
}
\value{
A Tibble
}
\description{
A wrapper around \code{googlesheets4::read_sheet()} to read all sheets from a spread sheet.
Output will be a list of data.frame with names corresponding to Google sheets (tabs) names.
}
